refactor(github): split 600-line monolith into focused single-responsibility classes
- Split GithubClient into 4 focused classes:
  * GraphQLClient (247 lines): Handles GraphQL operations
  * RESTClient (254 lines): Handles REST API operations
  * GithubClient (165 lines): Main orchestrator
  * Utils (67 lines): Shared utility functions
- Each class now has single responsibility and clear interface
- Improved testability with focused classes
- Better separation of GraphQL vs REST concerns
- Functionality verified: application works correctly with new architecture
